ETL/Producers/live_price_producer/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
pika==1.3.2
requests==2.32.3
urllib3==2.2.3
websocket-client==1.8.0
pymongo==4.10.1

